Commit_14

+++ AccountSystemLogic/src/main/java/za/ac/nwu/as/logic/flow/FetchAccountTransactionFlow.java
@@ -1,11 +1,10 @@
 package za.ac.nwu.as.logic.flow;
 
 import za.ac.nwu.as.domain.dto.AccountTransactionDto;
-
 import java.util.List;
 
 public interface FetchAccountTransactionFlow {
     List<AccountTransactionDto> getAllAccountTransactions();
 
-    AccountTransactionDto getAccountTransactionById(Long transactionId);
+    AccountTransactionDto getAccountTransactionById(int transactionId);
 }
